the module NumberClass.py did not conform to pep naming conventions, and so was renamed to number_class.py
