+++ msspackages/parity/__init__.py
@@ -1,1 +1,6 @@
-from .NumberClass import Number
+"""
+Example class for purposes of demonstrating contributions to MSSPackages
+"""
+
+
+from .number_class import Number
